Fix minimizing not hiding to taskbar
Prior to this change, whenever the application minimized, it would
completely delete the GUI and the taskbar icon. This meant that the only
way to unminimize was to run the app again.

This change comments out the code causing unexpected behaviour and adds
ICONIFIED JFrame state to create expected behaviour.