Add model constructors & edit parameters as needed
Prior to this change, none of the models had constructors. This made the
use of the models very difficult when you have to read and write them.
Additionally, Route did not contain any nodes and also stored length and
duration information as variables, which is incorrect. Those should be
implemented in US3 as calculations.

This change creates constructors for every model and edits Route to make
it correct.